2008-12-14	Joel Sherrill <joel.sherrill@oarcorp.com>
	* libcsupport/src/realloc.c, rtems/include/rtems/rtems/region.h,
	rtems/src/regioncreate.c, rtems/src/regionextend.c,
	rtems/src/regiongetsegment.c, rtems/src/regiongetsegmentsize.c,
	rtems/src/regionmp.c, rtems/src/regionresizesegment.c,
	sapi/include/rtems/config.h, score/include/rtems/score/heap.h,
	score/include/rtems/score/protectedheap.h, score/src/heap.c,
	score/src/heapallocate.c, score/src/heapallocatealigned.c,
	score/src/heapextend.c, score/src/heapresizeblock.c,
	score/src/heapsizeofuserarea.c, score/src/pheapallocate.c,
	score/src/pheapallocatealigned.c, score/src/pheapextend.c,
	score/src/pheapgetblocksize.c, score/src/pheapresizeblock.c: Change
	sizes of heap/region and allocated objects in heap to intptr_t so
	they can be larger than a single allocatable object (e.g. size_t).

+++ cpukit/rtems/src/regioncreate.c
@@ -50,7 +50,7 @@
 rtems_status_code rtems_region_create(
   rtems_name          name,
   void               *starting_address,
-  uint32_t            length,
+  intptr_t            length,
   uint32_t            page_size,
   rtems_attribute     attribute_set,
   Objects_Id         *id
