+++ cpukit/score/src/heapextend.c
@@ -38,8 +38,8 @@
 Heap_Extend_status _Heap_Extend(
   Heap_Control        *the_heap,
   void                *starting_address,
-  size_t               size,
-  uint32_t            *amount_extended
+  intptr_t             size,
+  intptr_t            *amount_extended
 )
 {
   uint32_t         the_size;
