+++ cpukit/score/src/pheapresizeblock.c
@@ -19,12 +19,12 @@
 bool _Protected_heap_Resize_block(
   Heap_Control *the_heap,
   void         *starting_address,
-  size_t        size
+  intptr_t      size
 )
 {
   Heap_Resize_status status;
-  uint32_t           old_mem_size;
-  uint32_t           avail_mem_size;
+  intptr_t           old_mem_size;
+  intptr_t           avail_mem_size;
 
   _RTEMS_Lock_allocator();
     status = _Heap_Resize_block(
